Split workflow_gantt tests
